refactor: tryCount를 Player 내부로 이동
