+++ src/main/java/bridge/domain/Player.java
@@ -1,13 +1,23 @@
 package bridge.domain;
 
 public class Player {
-    private int position;
+    private int tryCount = 1;
+    private int position = 0;
 
     public void moveForward() {
         position++;
     }
 
+    public void addTryCount() {
+        this.position = 0;
+        tryCount++;
+    }
+
     public int getPosition() {
         return position;
+    }
+
+    public int getTryCount() {
+        return tryCount;
     }
 }
