--- src/main/java/bridge/domain/TryCount.java
@@ -1,13 +1,0 @@
-package bridge.domain;
-
-public class TryCount {
-    private int tryCount = 1;
-
-    public void addTryCount() {
-        tryCount++;
-    }
-
-    public int getTryCount() {
-        return tryCount;
-    }
-}
